Create an Object so that works can be referred to by key
Not used yet, but the goal is eventually to make the song data array something like
`memberOf: new Set([AKYU5, HRTP])` or something instead of the cursed unary encoding it is
right now.

This change has no functional change. It was verified by doing the following:

function checkTitles() {
	if (ary_TitleData.length !== ary_TitleData2.length) {
		return false;
	}

	for (let i = 0; i < ary_TitleData.length; i++) {
		if (ary_TitleData[i] !== ary_TitleData2[i]) {
			return false;
		}
	}

	return true;
}

where `TitleData2` was the title array after this diff, and TitleData was the existing
array.